Add compat.py Move configparser, and urllib.parse imports here Move basestring definition here Update imports
