+++ praw/models/listing/mixins/base.py
@@ -1,5 +1,5 @@
 """Provide the BaseListingMixin class."""
-from ....const import urljoin
+from ....compat import urljoin
 from ...base import PRAWBase
 from ..generator import ListingGenerator
