+++ praw/models/listing/mixins/gilded.py
@@ -1,5 +1,5 @@
 """Provide the GildedListingMixin class."""
-from ....const import urljoin
+from ....compat import urljoin
 from ...base import PRAWBase
 from ..generator import ListingGenerator
