+++ tests/unit/test_reddit.py
@@ -3,8 +3,8 @@ import types
 import mock
 import pytest
 from praw import __version__, Reddit
+from praw.compat import configparser
 from praw.config import Config
-from praw.const import configparser
 from praw.exceptions import ClientException
 from prawcore import Requestor
